pyproject.toml
[project]
name = "owlv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "easygui>=0.98.3",
    "gradio>=5.34.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scipy>=1.15.3",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
